+++ file.go
@@ -1,12 +1,16 @@
 package simple
 
 import (
-	"encoding/json"
-	"encoding/xml"
-	"fmt"
+	"io"
 	"os"
+	"fmt"
 	"path"
+	"context"
 	"strings"
+	"archive/zip"
+	"encoding/xml"
+	"path/filepath"
+	"encoding/json"
 
 	"github.com/go-playground/validator/v10"
 	"github.com/pelletier/go-toml"
@@ -22,6 +26,7 @@ var FileRegistry = map[string]FileParser{
 	"xml":  xml.Unmarshal,
 }
 
+// Deserialize data from the given data type (json, yaml, toml or xml)
 func LoadFile[S any](filePath string, limit int, validation bool) ([]S, error) {
 	data, err := os.ReadFile(filePath)
 	if err != nil {
@@ -57,4 +62,67 @@ func LoadFile[S any](filePath string, limit int, validation bool) ([]S, error) {
 	}
 
 	return items, nil
+}
+
+// Adapted from Gosamples website
+func Unzip(ctx context.Context, source, destination string) error {
+    reader, err := zip.OpenReader(source)
+    if err != nil {
+        return err
+    }
+    defer reader.Close()
+
+    destination, err = filepath.Abs(destination)
+    if err != nil {
+        return err
+    }
+
+    for _, f := range reader.File {
+        select {
+			case <-ctx.Done():
+				return ctx.Err()
+			default:
+				err := unzipFile(f, destination)
+        		if err != nil {
+            		return err
+        		}
+        }
+    }
+    return nil
+}
+
+func unzipFile(f *zip.File, destination string) error {
+    // Check if file paths are not vulnerable to Zip Slip
+    filePath := filepath.Join(destination, f.Name)
+    if !strings.HasPrefix(filePath, filepath.Clean(destination)+string(os.PathSeparator)) {
+        return fmt.Errorf("invalid file path: %s", filePath)
+    }
+
+    if f.FileInfo().IsDir() {
+        if err := os.MkdirAll(filePath, os.ModePerm); err != nil {
+            return err
+        }
+        return nil
+    }
+
+    if err := os.MkdirAll(filepath.Dir(filePath), os.ModePerm); err != nil {
+        return err
+    }
+
+    destinationFile, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
+    if err != nil {
+        return err
+    }
+    defer destinationFile.Close()
+
+    zippedFile, err := f.Open()
+    if err != nil {
+        return err
+    }
+    defer zippedFile.Close()
+
+    if _, err := io.Copy(destinationFile, zippedFile); err != nil {
+        return err
+    }
+    return nil
 }